doc(xucred): remove PID lookup recommendation on FreeBSD
This actually isn't a good idea because, as explained in commit
5810829325b8eee05fa4bd67b7e9cc48e5797e3e, the credentials returned are
from the time that connect()/socketpair() was called. If that process
died and the PID has since been reused, looking up the PID to determine
the supplementary group list could result in security issues unless
appropriate caution is taken.